+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added logging for getRates
+- Added time spent metrics for getRates
+
 ## [1.2.0] - 2022-03-25
 
 ### Added